Organized the first completed demonstration into its own group
Xcode provides no easy way to create groups (what xcode calls folders)
that exist as folders on hard disk. To put all the files associate with
the first completed demonstration in a sub-folder of Classes I had to
follow these steps:

1: Create a new folder in Finder
2: Create a new group in project using xcode
3: Using Finder, move the files to their new folder
4: in Xcode, these files should now appear red. Drag and drop them into
the group created in step 2
5: In Xcode, for each moved file use the 'File Inspector' tab to change
its path to point at the file's location on the hard disk. (the button
to select a new path looks like a window and is directly beneath the
'Path' dropdown menu